Updated credits and added temporary reference directory containing Departure Board.
See TempDepartureBoard folder for inspiration/code for setting up a GUI. This folder must be purged eventually, but important that appropriate credits are kept (kryfield) - already added to strings, README and mod.lua.

With the addition of a custom GUI, I believe the mod.lua param settings can be killed - most of the settings will not only be load-time stuff, but things we want to toggle in-game whilst the game is running.